jobsy/storage/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
boto3>=1.34.0
Pillow>=10.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
